change multi, deleted, blocked to bigint

+++ packages/migration/ecosystem.go
@@ -959,9 +959,9 @@ MenuItem(
 		"id" bigint  NOT NULL DEFAULT '0',
 		"pub" bytea  NOT NULL DEFAULT '',
 		"amount" decimal(30) NOT NULL DEFAULT '0',
-		"multi" int NOT NULL DEFAULT '0',
-		"deleted" int NOT NULL DEFAULT '0',
-		"blocked" int NOT NULL DEFAULT '0'
+		"multi" bigint NOT NULL DEFAULT '0',
+		"deleted" bigint NOT NULL DEFAULT '0',
+		"blocked" bigint NOT NULL DEFAULT '0'
 		);
 		ALTER TABLE ONLY "%[1]d_keys" ADD CONSTRAINT "%[1]d_keys_pkey" PRIMARY KEY (id);
